Enable Parsing of auto within lambdas - i.e. allow [](auto a, decltype(a) b) { }
Currently arrays (i.e auto (&a)[5)) are not working

The approach involves adding a flag in the declarator that if set while parsing a lambda expr (it is if generic lambda support is enabled) enables GetDeclSpecTypeForDeclarator to check for this flag before reporting an error while in the FunctionPrototypeContext.

The tests are not python testable just yet.