Allow use of decltype to constrain subsequent generic lambda parameters
 Check if this param references a prior parameter.
(i.e.decltype (a) b) and rewire it to refer to the new invented
parameter - otherwise it will not be linked to the corresponding
template parameter during instantiation.

Since we can not be sure of how deeply nested the reference
might be, or whether we are referring to multiple previous params
we use a Visitor to comprehensively adjust all the pointers.

If we find any references to the prior parameters we use the new
renditions that we invented above (i.e. those not containing 'auto')
instead
i.e. [](auto a, auto* a2, decltype(a) b, X<sizeof(a)> x,
         Y<decltype(a)> y,  int c, Z<decltype(x)> z,
          auto (*b)(decltype(a), decltype(a2)) )

Some Questions:
  1) Is it safe to just reset the Decl within the DeclRefExpr
      of the old parameter to the new one?
  2) Is using the visitor the best way to do this?