Fix deduction for recursive functions
This uses code from Richard Smith's return type deduction branch on github, and modifies it slightly.

The approach is that each time a return statement is encountered within an instantiation of a generic lambda body, and ActOnCapScopeReturnStmt is called, we attempt to deduce the return type, and if it can be deduced (i.e. non-recursive case) then the Function's result type is adjusted to it.

Thus when in deduceClosureReturnType() we iterate through all the return statements and figure out their return types, when we encounter the recursive case, since the recursive case refers to the functionDecl that had its type adjusted by a non-recursive return, it's type can be successfully deduced.