Fix capture errors, and prevent assertion failure on captures
Apparently the following would compile:
int local = 4;
auto L = [](auto a) local;
and then error upon L(5);

This patch, removes the IsCapturedByClosure flag introduced in DeclRefExpr since it proved useless.  The old logic would allow illegal captures in generic lambdas to compile at definition time.

This patch essentially forwards everything to be handled by tryCaptureVar if it is not in an unevaluated context and is a generic lambda expression.

In addition, an initial attempt is made to convert all the tests to using the testing-infrastructure.