Get decltype to play nice with the conversion operator
In essence we want the following to work:
  auto L = [](auto a, decltype(a) b) a + b;
  int (*fp)(int, int) = L;

We need to add the parameters from the original call operator
to allow decltype to work in the conversion operator and
the lambda static invoker.
When we create parameters for our lambda static-invoker/and types
for our conversion operator,
the parameters that refer to previous parameters
don't get re-wired - and if we try to rewire them via
re-setting the DeclRefExpr using a visitor, it affects the
DeclRefExpr of the lambda call operator - i.e the DRE node does
not get cloned, and so is shared and so can not be tampered with.

Since static-invoker/function call operator parameter do not actually
get called (so to speak), we use
a crappy kludge to get this to pass through instantiation
i.e we instantiate the lambda call operator's params into
the instantiation scope of the static-invoker (special
casing AddressResolver to merge parent scopes when dealing
with the special static invoker - ugh!)

FVTODO: What would be nice is if I can figure out how to run
a TreeTransform and clone all the parameters during creation, with them
being completely rewired up to point to their correct
declarations.
Anyways, until then heed this warning:
This way should be shut! It was made by those who are dead!
And the dead should bloody keep it...