Some more fixes for decltype referring to parameters in nested lambdas
Prior to this patch, the following would cause an assertion failure:
auto L = [](auto b)
            [](auto a) -> decltype(a) a;

This occurred because when computing/transforming the New Type Source Info for the prototype of the function call operator, when the parameters were transformed, they were added to a new localinstantiationscope that would get popped off, and so when attempting to create the Lambda using the "builders" from Sema, we would need to transform the function parameter types that would then create new instantiated Declarations that would get added to the current instantiation scope, and used to construct our lambda expression - but then this could introduce an inconsistency since the decltype(a) above would refer to a different declaration than the one that was used to construct the lambda

The solution I use is somewhat kludgey, but appears to work.
Introduce a flag in Sema so that the transformers can check if they are transforming a lambdacalloperator prototype and if so, add the instantiations to the currentinstantiationscope.

In addition I also added a stack in sema to keep a track of whether we are in the state of transforming a lambda - currently I have no use for this - but have some vague sense it might prove useful as more bug reports come in - so I just added it - can always remove it later.