Allow nested dependent types as the return type
This patch replaces a crappy kludge that I had used in the beginning to transform the auto containing lambda parameter clause into an overload function call operator.  I should have used the TypeLocBuilder from the beginning - but I wasn't sure how to do that until I got around to studying the code in SemaTemplateInstantiate and the TransformFunctionProtoType code but then was too lazy to fix the kludge.  Unsurprisingly, the kludge was messing up the location of the result type, which would cause TransformNestedNameSpecifier to fail on an assertion and crash.

This patch fixes that and more importantly feels more robust.

It allows the following to not crash:
template<class T> struct defer {
  using type = T;
};

auto L = [](auto a) -> typename defer<decltype(a)>::type
                       { return a; };