Allow generic lambdas as default arguments in function templates
This patch computes the template depth correctly for generic lambdas that are default arguments for function templates i.e.

int fi(int) { return 0; }
template<class T> void foo(T fp1, T fp2 = [](auto a) a) {

}
foo(&fi); // ok

We do this, by first checking the scope if it is a lambda scope, which is contained within a function prototype scope which is within a templateparam scope.  I also introduced a n extra DeclContext argument in the Scope variant of getTemplateParameterDepth family, in order to bypass searching the Scope for a DeclContext, and just forwarding the Lambda itself.

Added some basic tests for this patch