Prevent crashing after errors emitted for forgetting to capture a var
Prior to this patch, the following would crash:

auto L = [](auto a)[](auto b) a + b;
L(4)(5);

Appropriate errors would be emitted regarding forgetting to capture 'a', but then a crash would occur on L(4)(5);

So we don't try and build the call expression if we had an error - this is probably a little too eager in the sense, we could probably tolerate this and compile on emitting diagnostics if no call was actually made, but for now this works.  No crash.

Thanks to Scott Prager for reporting this Bug.