Check getCurLambda() only if Function is a lambda callop.
Add a check in InstantiateFunctionDefinition to check that
the function being specialized is truly a lambda call operator.

I wonder if getCurLambda() can return true for some outer scope lambda if the function being specialized is not a lambda?? sounds very unlikely.
Eitherway - better to be safe! perhaps this would better be an assert.