Add a different version of calculating the unique common type   - iterate through each return stmt     - find the set of all its common expressions and types when unified against every other return stmt     - within that set, find the common type, given all the commontypes and modified-exprs for this return stmt   - use that unified-common-type for the return stmt to now find the common type against every other return stmt - and we should not discover any new common types - if so, an error should occur.
Repeat this for every other Return stmt.

Yuk - this is big-Oh of N2 - and can probably be optimized.

Additionally, to handle nullptrs within return stmts when they are unified with other ptrs, make sure an explicit c-cast is added to the implicit cast - so that the type of that cast can not be subverted or modified (such as they can be for nullptrs that are implicitly cast and only remember the implicit cast, in the context of return type deduction, when we deduce other common types from them)