Implement recursion if recursive call is within the first return stmt of a conditional.
For e.g.
auto f(int n) {
  return n > 10 ? 5 : foo(++n);
}

This is done by keeping a track of when we are either parsing or transforming a return stmt, if we encounter a conditional, track its middle expression - and clear it out once we are done processing its end expression - and when doing return type deduction - if there are no other returns (actually we should only be tracking the conditional, if no returns on the return stms stack exists - perhaps will fix this in the next commit) to do the deduction from then use any tracked middle expressions to deduce from.

One essentially has to have seen at least one middle expression before deducing the recursive return type.